backend/requirements.txt
fastapi
uvicorn
pydantic
requests
guardrails-ai